refactor: split client into sapphire-agent-api and sapphire-call crates
Extract the call client logic from src/call.rs into a shared
sapphire-agent-api crate, and introduce sapphire-call as a standalone
CLI binary. This enables distributing the client independently (e.g.
bundled in a VSCode extension) without server dependencies.

- Convert project to a Cargo workspace with three members
- sapphire-agent-api: shared client library (JSON-RPC, SSE, REPL)
- sapphire-call: standalone client binary (thin clap wrapper)
- sapphire-agent: retains `call` subcommand via re-export

Closes #9
